# docs/contrato/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ServiceDetail": {
        "properties": {
          "alarm_signs": {
            "items": {
              "type": "string"
            },
            "title": "Alarm Signs",
            "type": "array"
          },
          "audience": {
            "title": "Audience",
            "type": "string"
          },
          "documents": {
            "items": {
              "type": "string"
            },
            "title": "Documents",
            "type": "array"
          },
          "emergency_phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Emergency Phone"
          },
          "frequency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Frequency"
          },
          "how_to_schedule": {
            "title": "How To Schedule",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "link": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ServiceLinkOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "title": "Notes",
            "type": "array"
          },
          "preparation": {
            "items": {
              "type": "string"
            },
            "title": "Preparation",
            "type": "array"
          },
          "reminders": {
            "items": {
              "$ref": "#/components/schemas/ServiceReminderOut"
            },
            "title": "Reminders",
            "type": "array"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/ServiceSourceOut"
            },
            "title": "Sources",
            "type": "array"
          },
          "synonyms": {
            "items": {
              "type": "string"
            },
            "title": "Synonyms",
            "type": "array"
          },
          "where": {
            "title": "Where",
            "type": "string"
          },
          "who": {
            "title": "Who",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "audience",
          "synonyms",
          "who",
          "where",
          "how_to_schedule",
          "documents",
          "preparation",
          "alarm_signs",
          "notes",
          "reminders",
          "sources"
        ],
        "title": "ServiceDetail",
        "type": "object"
      },
      "ServiceLinkOut": {
        "properties": {
          "label": {
            "title": "Label",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "label",
          "url"
        ],
        "title": "ServiceLinkOut",
        "type": "object"
      },
      "ServiceReminderOut": {
        "properties": {
          "days": {
            "title": "Days",
            "type": "integer"
          },
          "label": {
            "title": "Label",
            "type": "string"
          }
        },
        "required": [
          "label",
          "days"
        ],
        "title": "ServiceReminderOut",
        "type": "object"
      },
      "ServiceSourceOut": {
        "properties": {
          "publisher": {
            "title": "Publisher",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "year": {
            "title": "Year",
            "type": "integer"
          }
        },
        "required": [
          "title",
          "publisher",
          "year"
        ],
        "title": "ServiceSourceOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "RagTest API",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/services/{service_id}": {
      "get": {
        "operationId": "get_service_v1_services__service_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "service_id",
            "required": true,
            "schema": {
              "title": "Service Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Service",
        "tags": [
          "catalog"
        ]
      }
    },
...
  }
}